+++ source/extensions/lightspeed.trex.stage_manager.plugin.interaction.usd/lightspeed/trex/stage_manager/plugin/interaction/usd/all_categories.py
@@ -41,6 +41,7 @@ class RemixAllCategoriesInteractionPlugin(RemixStageManagerUSDInteractionPlugin)
     compatible_filters: list[str] = Field(
         default=[
             "AdditionalFilterPlugin",
+            "CustomTagsFilterPlugin",
             "IgnorePrimsFilterPlugin",
             "IsCaptureFilterPlugin",
             "IsCategoryFilterPlugin",