+++ source/extensions/omni.flux.stage_manager.plugin.interaction.usd/omni/flux/stage_manager/plugin/interaction/usd/all_lights.py
@@ -43,6 +43,7 @@ class AllLightsInteractionPlugin(_StageManagerUSDInteractionPlugin):
     compatible_filters: list[str] = Field(
         default=[
             "AdditionalFilterPlugin",
+            "CustomTagsFilterPlugin",
             "IgnorePrimsFilterPlugin",
             "VisiblePrimsFilterPlugin",
             "LightPrimsFilterPlugin",