+++ source/extensions/omni.flux.stage_manager.plugin.interaction.usd/omni/flux/stage_manager/plugin/interaction/usd/all_prims.py
@@ -34,6 +34,7 @@ class AllPrimsInteractionPlugin(_StageManagerUSDInteractionPlugin):
     compatible_filters: list[str] = Field(
         default=[
             "AdditionalFilterPlugin",
+            "CustomTagsFilterPlugin",
             "GeometryPrimsFilterPlugin",
             "IgnorePrimsFilterPlugin",
             "VisiblePrimsFilterPlugin",